SERVER-43239 Fixed bug causing numCatchUpOps in repSetGetStatus to be incorrect
(cherry picked from commit 71e4779b0da9e8d58dbb179c49b1a86c5e48c93d)

SERVER-41512 Added tracking for metrics around a node voting in an election

(cherry picked from commit 7538504cb584720c2cbbc6d44ea62d0743b41fcf)

SERVER-41513 Track the time the new term oplog entry was written by primary and applied in secondary

(cherry picked from commit efde009845f32d8de2d094088628e67608bfa419)

+++ src/mongo/db/repl/replication_coordinator_mock.h
@@ -306,7 +306,7 @@ public:
 
     virtual Status abortCatchupIfNeeded(PrimaryCatchUpConclusionReason reason) override;
 
-    virtual void incrementNumCatchUpOpsIfCatchingUp(int numOps) override;
+    virtual void incrementNumCatchUpOpsIfCatchingUp(long numOps) override;
 
     void signalDropPendingCollectionsRemovedFromStorage() final;
